exp355 fix(runtime): uniq_protos 显式fp32(AMP下img_proj.new_zeros是Half, index_put赋fp32崩); 双审查漏的AMP index_put dtype

+++ model/pose_backbone_model.py
@@ -582,7 +582,7 @@ class PoseBackboneModel(build_transformer):
         P = uniq_labels.shape[0]
         if P < 3:
             return img_proj.new_zeros(())          # need >=2 hard-negatives for a soft target
-        uniq_protos = img_proj.new_zeros(P, txt_proto.shape[1])
+        uniq_protos = img_proj.new_zeros(P, txt_proto.shape[1], dtype=torch.float32)  # fp32: match txt_proto.float() (AMP index_put dtype)
         uniq_protos[inv] = txt_proto.float()
         # image-to-ID-prototype logits (fp32 for softmax stability under AMP)
         img_n = F.normalize(img_proj.float(), dim=1)
